+++ src/memoria.h
@@ -52,6 +52,7 @@ typedef enum {
     GC_TIPO_CLASE,
     GC_TIPO_INSTANCIA,
     GC_TIPO_METODO_LIGADO,
+    GC_TIPO_MODULO,
 } TipoGC;
 
 /*